pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intentforge"
version = "0.1.0"
description = "NLP-driven code generation framework via MQTT"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Softreck", email = "info@softreck.dev"}
]
maintainers = [
    {name = "Softreck", email = "info@softreck.dev"}
]
keywords = [
    "code-generation",
    "nlp",
    "mqtt",
    "llm",
    "api",
    "automation",
    "intent",
    "dsl"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "anthropic>=0.30.0",
    "paho-mqtt>=2.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "bandit>=1.7",
    "safety>=2.3",
]
test = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.2",
    "pytest-benchmark>=4.0",
    "httpx>=0.27",
    "respx>=0.20",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]
server = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
]
all = [
    "intentforge[dev,test,docs,server]",
]

[project.scripts]
intentforge = "intentforge.cli:main"
intentforge-server = "intentforge.server:run"

[project.urls]
Homepage = "https://github.com/wronai/intent"
Documentation = "https://intentforge.readthedocs.io"
Repository = "https://github.com/wronai/intent.git"
Changelog = "https://github.com/wronai/intent/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["intentforge*"]

[tool.setuptools.package-data]
intentforge = ["py.typed", "templates/*"]

# Ruff configuration
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "UP",   # pyupgrade
]
ignore = [
    "E402",   # module level import not at top of file
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "F841"]

[tool.ruff.lint.isort]
known-first-party = ["intentforge"]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["intentforge"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
